requirements.txt
Flask
spotipy
requests
python-dotenv
SQLAlchemy